+++ PyroEngine/src/PyroEngine/Renderer/VertexBuffer.h
@@ -11,6 +11,6 @@ namespace PyroEngine
 		virtual const VertexBufferLayout& GetLayout() const = 0;
 		virtual void SetLayout(const VertexBufferLayout& layout) = 0;
 
-		static VertexBuffer* Create(double* vertices, uint64_t size);
+		static VertexBuffer* Create(float* vertices, uint64_t size);
 	};
 }
